xmu-tronclass-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xmu-tronclass-sdk"
version = "0.1.0"
description = "Python SDK for TronClass LMS — supports all APK features including push-based rollcall"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "KrsMt", email = "krsmt0113@gmail.com"}
]
keywords = ["tronclass", "xmu", "xiamen-university", "lms", "rollcall", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Education",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28",
    "pycryptodome>=3.18",
    "aiohttp>=3.9",
    "websockets>=12.0",
    "python-socketio[asyncio_client]>=5.10",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.urls]
Repository = "https://github.com/KrsMt-0113/XMU-Tronclass-SDK"

[tool.setuptools.packages.find]
where = ["."]
include = ["tronclass*"]
